Create code
The provided code appears to be written in Arduino programming language, which is based on C/C++. It is used for programming Arduino microcontrollers. This code defines a relay using the `#define relay` directive and sets up a communication interface via Serial at a baud rate of 9600. It reads data from the Serial interface and controls the state of pin 2 based on the received input.